cleanup appendables for performance
added helper methods for REST calls

+++ src/main/java/com/ociweb/pronghorn/util/math/DecimalResult.java
@@ -2,6 +2,6 @@ package com.ociweb.pronghorn.util.math;
 
 public interface DecimalResult {
 
-	void result(long m, int e);
+	void result(long m, byte e);
 
 }
